backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
stripe
sentry-sdk
